Add Atomic Hosts as a generic driver
This adds support for Project Atomic hosts (projectatomic.io) which includes
Fedora Atomic as well as CentOS Atomic as a generic driver to Docker
Machine.

This relies on VARIANT_ID being set to "atomic.host" which has been
added to Fedora Atomic 24 as per
http://pkgs.fedoraproject.org/cgit/rpms/fedora-release.git/commit/?h=f24&id=cc3e5a1d378dcdf0294325e22d060de7441c933f
and in the future to CentOS Atomic.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>